From kolya: pass pointer to int, not long, to ioctl.
git-svn-id: svn+ssh://svn.mit.edu/athena/trunk@15093 728af825-273c-0410-89f9-f7d3b574a069

+++ third/xmh/command.c
@@ -390,7 +390,7 @@ CheckReadFromPipe( fd, bufP, lenP, waitEOF )
     int *lenP;
     Bool waitEOF;
 {
-    long nread;
+    int nread;
 /*  DEBUG2( " CheckReadFromPipe #%d,len=%d,", fd, *lenP )  */
 #ifdef FIONREAD
     if (!ioctl( fd, FIONREAD, &nread )) {
